server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.4.0
rembg==2.0.55
onnxruntime==1.16.3
